requirements.txt
streamlit
yfinance
fintoolkit
xlrd
openpyxl
redis